requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0

# Machine Learning
scikit-learn>=1.0.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
